+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project follows [semantic versioning](http://semver.org).
 
 
 ### UNRELEASED
+* [added] Ability to read/write tar file from/to stdin/stdout
+* [modified] Logging to stderr
 * [fixed] Only print "repairing bundles" if actually repairing bundles